requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.2
aiosqlite==0.20.0
annotated-types==0.7.0
anthropic==0.42.0
anyio==4.7.0
attrs==24.3.0
certifi==2024.12.14
discord.py==2.4.0
distro==1.9.0
docstring_parser==0.16
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.8.2
mirascope==1.13.1
multidict==6.1.0
openai==1.58.1
propcache==0.2.1
pydantic==2.10.3
pydantic_core==2.27.1
python-dotenv==1.0.1
PyYAML==6.0.2
sniffio==1.3.1
tqdm==4.67.1
typing_extensions==4.12.2
yarl==1.18.3
